Save the CMPIBroker in the Cmpi module
The CMPIBroker is required for some callbacks, e.g. when creating a
CMPIString.

This change passes the broker value from the provider initialization
to the Cmpi module which in saves it as module-level variable.

This make it possible to retrieve the broker value inside the bindings
since the module pointer is available as 'mCmpi'