Replace to_s() with __str__() in Python
Definition of to_s() was a Ruby-ish leftover, esp for
CMPIStatus.to_s() which required a CMPIBroker to be passed.

Now that we have a global CMPIBroker, CMPIStatus.to_s() in Python is
deprecated but left in place for backwards-compatibility.

All other CMPI classes should have proper __str__() functions defined
in Python (resp. to_s() in Ruby) now.